requirements.txt
playwright==1.49.1
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
Pillow==11.0.0
lxml==5.2.2
reportlab==4.0.6
